requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2024.1.1
openai-whisper==20231117
python-multipart==0.0.6
youtube-transcript-api==0.6.7
